+++ src/Interfaces/WordPress/Posts/GetMetaDataInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Posts;
+
+interface GetMetaDataInterface
+{
+    public function getMetaData(): array;
+}