+++ src/Interfaces/WordPress/Posts/RegisterCustomTypeInterface.php
@@ -5,7 +5,7 @@ namespace ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Posts;
 /**
  * Let's keep those custom post types all in line.
  */
-interface CustomTypeRegisterInterface
+interface RegisterCustomTypeInterface
 {
     public function registerCustomType(): void;
 }